Fix greedy manager bug and other hidden bugs too.
We only checked if more jobs should be fetched when we pushed a new job
to the worker. That means that if after pushing the first job we decided
we shouldn't fetch any more jobs (because that job is pretty big), we
would never check again.

The correct way is to timely check if more jobs are needed. We also use
a "huge" number as cost prediction when a prediction is not available.
This simplifies the logic (as opposed to using 0 as no prediction
value).